app/ProviderAgent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ProviderAgent"
version = "0.1.0"
description = "Healthcare provider lookup agent on AWS Bedrock AgentCore"
requires-python = ">=3.10"
dependencies = [
    "bedrock-agentcore >= 1.0.3",
    "strands-agents >= 1.13.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]
